Support multiple batteries for the battery module
Changes the calculation of the battery percentage to take into account
many batteries. This is by summing their current energy & capacity, then
getting a percentage from that. This is what the `battery` crate
describes what `Battery::state_of_charge` does but with less accuracy.
This should be fine because the percentage is just 2 s.f.

Since there is now many batteries the overall state of the battery must
be merged from many to one. The strategy here is: if any are discharging
then report as discharging; if none are discharging & at least one is
charging, then report as charging; if the battery states (e.g. full/empty)
are the same, report as that; otherwise, report as unknown.

Closes issue #656